auto-life/requirements.txt
opencv_python==4.1.1.26
PyMySQL==0.9.3
selenium==3.141.0
xpinyin==0.5.6
numpy==1.17.3
PyUserInput==0.1.11
xlrd==1.2.0
Appium_Python_Client==1.0.2
Pillow==7.2.0
pymouse==1.0
website==1.1
xlsxwriter==1.3.3
